Add folder
Hi Everett, Joe here! This is my first attempt at a pull request. I noticed that one has to download it, but the download link isn't permanent. To that end, I am adding the folder you created to make it easier for others to use. Lemme know what you think.